+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * **assessment** node response saving module now also accepts **templateResponse** as input
 * **startSession** node will use database TZ for newly inserted session datetime values
 * **finishSession** node will use database TZ for returned session datetime values
+* exposed startTestTimer and stopTestTimer js testRunner methods
 
 #### Docker